Move reusable components to 'shared' directory (#134)
* Move reusable components to 'shared' directory.
* Make DashboardHeader component reusable.
