+++ src/components/shared/EuroCurrency.tsx
